requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
xgboost==2.0.3
scikit-learn==1.4.2
nba_api==1.4.1
apscheduler==3.10.4
aiofiles==23.2.1
